RF FreestyleProjectTest > Combining tests
